requirements.txt
asgiref==3.3.1
boto3==1.16.43
botocore==1.19.43
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
django-rest-swagger==2.1.2
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==2.0.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
